move accordion and alert stories into component folderes
