+++ packages/react-components/react-accordion/src/stories/Accordion/AccordionExpandIconPosition.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { AccordionItem, AccordionHeader, AccordionPanel, Accordion } from '../index';
+import { AccordionItem, AccordionHeader, AccordionPanel, Accordion } from '../../index';
 
 export const ExpandIconPosition = () => (
   <Accordion>
