+++ packages/react-components/react-accordion/src/stories/Accordion/AccordionInline.stories.tsx
@@ -1,24 +1,22 @@
 import * as React from 'react';
-import { RocketRegular } from '@fluentui/react-icons';
+import { AccordionItem, AccordionHeader, AccordionPanel, Accordion } from '../../index';
 
-import { AccordionItem, AccordionHeader, AccordionPanel, Accordion } from '../index';
-
-export const WithIcon = () => (
+export const Inline = () => (
   <Accordion>
     <AccordionItem value="1">
-      <AccordionHeader icon={<RocketRegular />}>Accordion Header 1</AccordionHeader>
+      <AccordionHeader inline>Accordion Header 1</AccordionHeader>
       <AccordionPanel>
         <div>Accordion Panel 1</div>
       </AccordionPanel>
     </AccordionItem>
     <AccordionItem value="2">
-      <AccordionHeader icon={<RocketRegular />}>Accordion Header 2</AccordionHeader>
+      <AccordionHeader inline>Accordion Header 2</AccordionHeader>
       <AccordionPanel>
         <div>Accordion Panel 2</div>
       </AccordionPanel>
     </AccordionItem>
     <AccordionItem value="3">
-      <AccordionHeader icon={<RocketRegular />}>Accordion Header 3</AccordionHeader>
+      <AccordionHeader inline>Accordion Header 3</AccordionHeader>
       <AccordionPanel>
         <div>Accordion Panel 3</div>
       </AccordionPanel>
@@ -26,10 +24,10 @@ export const WithIcon = () => (
   </Accordion>
 );
 
-WithIcon.parameters = {
+Inline.parameters = {
   docs: {
     description: {
-      story: 'An accordion header can contain an icon.',
+      story: 'An accordion header can be set to `inline`',
     },
   },
 };
