+++ packages/react-components/react-accordion/src/stories/Accordion/AccordionNavigation.stories.tsx
@@ -1,33 +1,33 @@
 import * as React from 'react';
-import { AccordionItem, AccordionHeader, AccordionPanel, Accordion } from '../index';
+import { AccordionItem, AccordionHeader, AccordionPanel, Accordion } from '../../index';
 
-export const Disabled = () => (
-  <Accordion>
-    <AccordionItem disabled value="1">
+export const Navigation = () => (
+  <Accordion navigation="linear">
+    <AccordionItem value="1">
       <AccordionHeader>Accordion Header 1</AccordionHeader>
       <AccordionPanel>
-        <div>Accordion Panel 1</div>
+        <button>Accordion Panel 1</button>
       </AccordionPanel>
     </AccordionItem>
-    <AccordionItem disabled value="2">
+    <AccordionItem value="2">
       <AccordionHeader>Accordion Header 2</AccordionHeader>
       <AccordionPanel>
-        <div>Accordion Panel 2</div>
+        <button>Accordion Panel 2</button>
       </AccordionPanel>
     </AccordionItem>
-    <AccordionItem disabled value="3">
+    <AccordionItem value="3">
       <AccordionHeader>Accordion Header 3</AccordionHeader>
       <AccordionPanel>
-        <div>Accordion Panel 3</div>
+        <button>Accordion Panel 3</button>
       </AccordionPanel>
     </AccordionItem>
   </Accordion>
 );
 
-Disabled.parameters = {
+Navigation.parameters = {
   docs: {
     description: {
-      story: 'An accordion item can be `disabled`',
+      story: 'An accordion supports keyboard navigation.',
     },
   },
 };
