+++ packages/react-components/react-alert/src/stories/Alert/AlertDefault.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
 
-import { Alert } from '../index';
+import { Alert } from '../../index';
 
 export const Default = () => <Alert>This is a default alert</Alert>;
