+++ packages/react-components/react-alert/src/stories/Alert/AlertIcon.stories.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 
 import { DeleteRegular } from '@fluentui/react-icons';
 
-import { Alert } from '../index';
+import { Alert } from '../../index';
 
 export const Icon = () => (
   <Alert
