+++ packages/react-components/react-alert/src/stories/Alert/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Alert } from '../index';
+import { Alert } from '../../index';
 import descriptionMd from './AlertDescription.md';
 
 export { Default } from './AlertDefault.stories';
